% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/add_address.R
\name{add_address}
\alias{add_address}
\title{Include a formatted address in govspeak output}
\usage{
add_address(text)
}
\arguments{
\item{text}{address as a string}
}
\description{
Adds conditional address tags in govspeak only. Displays as formatted text in govspeak, and as simple text only in other formats.
}
