% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/add_blue_header.R
\name{add_blue_header}
\alias{add_blue_header}
\title{Add the blue header bar at the top of a HTML pre-release}
\usage{
add_blue_header(
  header_text,
  stats_status = "National Statistics",
  pre_release_note = "Pre-release"
)
}
\arguments{
\item{header_text}{a string, containing the title of the pre-release to include in the header text}

\item{stats_status}{a string, designating the statistics status of the pre-release. Defaults to "National Statistics", but can be Official or Experimental statistics}

\item{pre-release_note}{a string, containing  a pre-release note that will appear underneath the title in the header text. Defaults to "Pre-release".}
}
\description{
Add a blue header bar to the top of a HTML pre-release containing the stats release details to mimic the appearance of gov.uk
This should be done for the pre-release only, and not for the final statistics publication
}
