% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/add_email.R
\name{add_email}
\alias{add_email}
\title{Include an email address as a link}
\usage{
add_email(text)
}
\arguments{
\item{text}{email address as string}
}
\description{
Adds linked email address in govspeak only. Displays as a link in govspeak, and as simple text only in other formats.
}
