% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/add_logo.R
\name{add_logo}
\alias{add_logo}
\title{Add a logo to the top of a HTML pre-release}
\usage{
add_logo(img)
}
\arguments{
\item{img}{string giving the location of the logo as a png file.}
}
\description{
Add a logo to the top of a HTML pre-release to mimic the appearance of gov.uk
This should be done for the pre-release only, and not for the final
statistics publication
}
