% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/internal_functions.R
\name{bold_text}
\alias{bold_text}
\title{Add RMarkdown annotation to bold all text in a vector}
\usage{
bold_text(data)
}
\arguments{
\item{data}{vector containing strings of interest}
}
\description{
Bold all text in a vector
}
\keyword{internal}
