% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/callout_box.R
\name{callout_box}
\alias{callout_box}
\title{Include a callout box in govspeak output}
\usage{
callout_box(text)
}
\arguments{
\item{text}{string to publish in a callout box}
}
\description{
Adds conditional CTA tags in govspeak only. Displays as text in a box in govspeak, and as simple text only in other formats.
}
