% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/conditional_publishing_output.R
\name{conditional_publishing_output}
\alias{conditional_publishing_output}
\title{Conditional display of object depending on Rmarkdown publication type.}
\usage{
conditional_publishing_output(publication_type, output)
}
\arguments{
\item{publication_type}{string describing type of RMarkdown output you would like the object to appear in. Options include "html" for html_document, "docx" for word_document, and "pdf" for pdf_document.}

\item{output}{object to display in the document; can be a table, text, graph, or image.}
}
\description{
Conditional display of specified objects when RMarkdown document is knitted.
Expression will display the object only in the specified Rmarkdown output type.
}
