% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/internal_functions.R
\name{generate_image_references}
\alias{generate_image_references}
\title{Generate govdown image references}
\usage{
generate_image_references(lines)
}
\arguments{
\item{lines}{raw contents of md file (from readLines)}
}
\description{
Extracts image references from the original markdown output into Whitehall format
}
\keyword{internal}
