% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/knitr_opts_set.R
\name{knitr_opts_set}
\alias{knitr_opts_set}
\title{Set knitr parameters to allow publishing on Whitehall}
\usage{
knitr_opts_set(fig.path = "graphs/")
}
\arguments{
\item{fig.path}{folder to save generated images to. Defaults to "graphs"}
}
\description{
Set parameters necessary for publishing on Whitehall.
Includes setting specific figure height and width and
suppressing warning messages
}
