% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/internal_functions.R
\name{order_img_ref}
\alias{order_img_ref}
\title{Order image references in a logical way}
\usage{
order_img_ref(img_ref, img_folder)
}
\arguments{
\item{img_ref}{table of image references}

\item{img_wd}{working directory that image paths are listed from.
In most cases, this will be the same folder as the md document is stored in}
}
\description{
Order image references in a logical way
}
\keyword{internal}
