% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/HTML_functions.R
\name{add_image}
\alias{add_image}
\title{Highlight in bold key words in summary text}
\usage{
add_image(file_name, folder = "graphs")
}
\arguments{
\item{file_name}{string giving the file name of the image only}

\item{folder}{string giving the directory the image is stored in, defaults to "graphs".}
}
\description{
Add images to both HTML and other formats
}
