% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/add_summary_table.R
\name{add_summary_table}
\alias{add_summary_table}
\title{Produce summary table for key points in conditional format}
\usage{
add_summary_table(left, middle, right, key_words = NULL)
}
\arguments{
\item{left}{vector of key points which will appear in left column}

\item{middle}{vector of arrow image calls which will appear in middle column}

\item{right}{vector of descriptions of key points which will appear in right column}

\item{key_words}{vector of key change words to highlight in right column. A preloaded vector of these key words is already included, but additional words can be added as desired.}
}
\description{
Add summary table containing key points of statistical release.
Traditionally these consist of key points, images of arrows to indicate direction of change, and detailed description of the key points.
This normal structure is outputted in word and pdf formats, while the govspeak format includes only the two text columns.
The govspeak format also highlights key words indicating direction of change in the description.
}
