% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/external_functions.R
\name{add_summary_table}
\alias{add_summary_table}
\title{Produce summary table for key points in conditional format}
\usage{
add_summary_table(left, middle, right)
}
\arguments{
\item{left}{vector of left column key points}

\item{middle}{vector of arrow image calls}

\item{right}{vector of right column descriptions of key points}
}
\description{
Add summary table. Contains arrows indicating direction in word and PDF format, contains only text with key words highlighted in govspeak format.
}
