% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/add_table.R
\name{add_table}
\alias{add_table}
\title{Format data in a table suitable for publication}
\usage{
add_table(data)
}
\arguments{
\item{data}{dataframe; dataframe of any size to be displayed as a table}
}
\description{
Add a table in Rmarkdown suitable for publication in govspeak and in word/PDF
}
