% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/external_functions.R
\name{add_table}
\alias{add_table}
\title{Format data in a table suitable for publication}
\usage{
add_table(data)
}
\arguments{
\item{data}{dataframe you want to display as a table}
}
\description{
Add a table suitable for publication in govspeak and in word/PDF
}
