% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/internal_functions.R
\name{bold_key_words}
\alias{bold_key_words}
\title{Highlight in bold key words in summary text}
\usage{
bold_key_words(data, key_words = NULL, key_words_remove = NULL)
}
\arguments{
\item{data}{character vector of summary points}

\item{key_words}{character vector of additional words you would like
to highlight in bold (defaults to NULL)}

\item{key_words_remove}{character vector of standard key words you would NOT
like to highlight in bold (defaults to NULL)}
}
\description{
Compares summary points to vector of words indicating change,
 and highlights relevant words in bold.
}
