% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/external_functions.R
\name{callout_box}
\alias{callout_box}
\title{Include a callout box in govspeak output}
\usage{
callout_box(text)
}
\arguments{
\item{text}{string you want to publish in a call-out box}
}
\description{
Include text in a callout box in govspeak only, shows simple text only in other formats.
}
