% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/HTML_functions.R
\name{conditional_publishing_output}
\alias{conditional_publishing_output}
\title{Highlight in bold key words in summary text}
\usage{
conditional_publishing_output(publication_type, output)
}
\arguments{
\item{publication_type}{string describing type of RMarkdown output you would like the output to appear in}

\item{output}{object you want to output into document; can be a table, text, graph, or image.}
}
\description{
Conditional display of specified outputs when RMarkdown document is knitted
}
