% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/convert_rmd.R
\name{convert_rmd}
\alias{convert_rmd}
\title{Convert standard markdown file to govspeak}
\usage{
convert_rmd(
  path,
  images_folder = "graphs",
  remove_blocks = FALSE,
  sub_pattern = TRUE
)
}
\arguments{
\item{path}{string; filename (including path) to *.md file for conversion}

\item{images_folder}{string; folder containing images for *.md file. Defaults
to "graphs"}

\item{remove_blocks}{bool; decision to remove block elements from *.md file.
This includes code blocks and warnings}

\item{sub_pattern}{bool or vector; decision to increase hashed headers by one level in *.md file.
TRUE will substitute all, FALSE will substitute none, while a vector of the desired substitution levels allows individual headings to be modified as required.
e.g. "#" will modify only first level headers.}
}
\description{
Convert markdown file to Whitehall Publisher (GOV.UK) govspeak markdown format
}
