% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/external_functions.R
\name{convert_rmd}
\alias{convert_rmd}
\title{Convert standard markdown file to govspeak}
\usage{
convert_rmd(path, images_folder = "graphs", remove_blocks = FALSE)
}
\arguments{
\item{path}{string; filename (including path) to *.md file for conversion}

\item{images_folder}{string; folder containing images for *.md file. Defaults
to "graphs"}

\item{remove_blocks}{bool; decision to remove block elements from *.md file.
This includes code blocks and warnings}
}
\description{
Convert markdown file to Whitehall Publisher (GOV.UK) govspeak markdown format
}
